feat: Implement thread creation flow (#31)
There is a backend mutation that creates a thread and another that streams a message asynchronously.

But the create thread is not yet implemented. This commit implements it.

# Flow for new chat:
1. The user is on the homepage (meaning this is a new chat)
2. They type some message in the input box
3. They press send
4. At this point we should flip on a loading state (call it isCreatingThread) and then do the mutation to create a thread
5. When that mutation succeeds we should immediately call the async stream message mutation optimistically, after which navigate to that page with the thread ID. And turn off that loading state and reset the message input box

While isCreatingThread is true the app should grey out the input box first (with the user's message inside) and show an animated loading spinner icon in place of the arrow icon in the send button.